requirements.txt
fastapi
uvicorn[standard]
chromadb
pydantic
PyPDF2
google-generativeai
python-dotenv
